refactored builtin tools to be a bit more organized
